+++ src/Drivers/Imagick/Driver.php
@@ -80,12 +80,9 @@ class Driver extends AbstractDriver
      *
      * @see DriverInterface::createAnimation()
      */
-    public function createAnimation(callable $init): ImageInterface
+    public function createAnimation(callable $animation): ImageInterface
     {
-        $animation = new AnimationFactory($this);
-        $init($animation);
-
-        return call_user_func($animation);
+        return AnimationFactory::build($this, $animation);
     }
 
     /**